backend/requirements.dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.dev.txt requirements.base.in requirements.dev.in requirements.production.in
#
appnope==0.1.3
    # via ipython
asgiref==3.7.2
    # via django
asttokens==2.4.0
    # via stack-data
attrs==23.1.0
    # via curlylint
backcall==0.2.0
    # via ipython
black==23.9.1
    # via -r requirements.dev.in
boto3==1.28.53
    # via django-s3-storage
botocore==1.31.53
    # via
    #   boto3
    #   s3transfer
certifi==2023.7.22
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via
    #   black
    #   curlylint
curlylint==0.13.1
    # via -r requirements.dev.in
decorator==5.1.1
    # via ipython
dj-database-url==2.1.0
    # via -r requirements.production.in
django==4.2.5
    # via
    #   -r requirements.base.in
    #   dj-database-url
    #   django-anymail
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-lifecycle
    #   django-s3-storage
    #   djangorestframework
django-anymail==10.1
    # via -r requirements.production.in
django-cors-headers==4.2.0
    # via -r requirements.base.in
django-debug-toolbar==4.2.0
    # via -r requirements.dev.in
django-lifecycle==1.1.2
    # via -r requirements.base.in
django-s3-storage==0.14.0
    # via -r requirements.production.in
djangorestframework==3.14.0
    # via -r requirements.base.in
djhtml==3.0.6
    # via -r requirements.dev.in
executing==1.2.0
    # via stack-data
gunicorn==21.2.0
    # via -r requirements.production.in
hashids==1.3.1
    # via -r requirements.base.in
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
ipython==8.15.0
    # via -r requirements.base.in
isort==5.12.0
    # via -r requirements.dev.in
jedi==0.19.0
    # via ipython
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
matplotlib-inline==0.1.6
    # via ipython
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via
    #   black
    #   gunicorn
    #   pytest
parso==0.8.3
    # via jedi
parsy==1.1.0
    # via curlylint
pathspec==0.11.2
    # via
    #   black
    #   curlylint
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==10.0.1
    # via qrcode
platformdirs==3.10.0
    # via black
pluggy==1.3.0
    # via pytest
prompt-toolkit==3.0.39
    # via ipython
psycopg2-binary==2.9.7
    # via -r requirements.base.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.16.1
    # via ipython
pypng==0.20220715.0
    # via qrcode
pytest==7.4.2
    # via pytest-django
pytest-django==4.5.2
    # via -r requirements.dev.in
python-dateutil==2.8.2
    # via botocore
python-dotenv==1.0.0
    # via -r requirements.base.in
pytz==2023.3.post1
    # via djangorestframework
qrcode[pil]==7.4.2
    # via -r requirements.base.in
requests==2.31.0
    # via django-anymail
ruff==0.0.292
    # via -r requirements.dev.in
s3transfer==0.6.2
    # via boto3
sentry-sdk==1.31.0
    # via -r requirements.production.in
shortuuid==1.0.11
    # via -r requirements.base.in
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.2
    # via ipython
toml==0.10.2
    # via curlylint
traitlets==5.10.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.8.0
    # via
    #   dj-database-url
    #   qrcode
urllib3==1.26.16
    # via
    #   botocore
    #   django-anymail
    #   requests
    #   sentry-sdk
wcwidth==0.2.6
    # via prompt-toolkit
whitenoise==6.5.0
    # via -r requirements.production.in
